+++ clients/admin-ui/src/types/api/models/ExperienceConfigUpdate.ts
@@ -2,25 +2,63 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import type { ComponentType } from "./ComponentType";
-import type { DeliveryMechanism } from "./DeliveryMechanism";
+import type { BannerEnabled } from "./BannerEnabled";
 import type { PrivacyNoticeRegion } from "./PrivacyNoticeRegion";
 
 /**
- * Updating ExperienceConfig - requires regions to be patched specifically
+ * Updating ExperienceConfig. Note that component cannot be updated once its created
  */
 export type ExperienceConfigUpdate = {
-  acknowledgement_button_label?: string;
-  banner_title?: string;
-  banner_description?: string;
-  component?: ComponentType;
-  component_title?: string;
-  component_description?: string;
-  confirmation_button_label?: string;
-  delivery_mechanism?: DeliveryMechanism;
+  /**
+   * Overlay 'Accept button displayed on the Banner and Privacy Preferences' or Privacy Center 'Confirmation button label'
+   */
+  accept_button_label?: string;
+  /**
+   * Overlay 'Acknowledge button label for notice only banner'
+   */
+  acknowledge_button_label?: string;
+  /**
+   * Overlay 'Banner'
+   */
+  banner_enabled?: BannerEnabled;
+  /**
+   * Overlay 'Banner Description' or Privacy Center 'Description'
+   */
+  description?: string;
+  /**
+   * Whether the given ExperienceConfig is disabled
+   */
   disabled?: boolean;
+  /**
+   * Whether the given ExperienceConfig is a global default
+   */
   is_default?: boolean;
-  link_label?: string;
+  /**
+   * Overlay and Privacy Center 'Privacy policy link label'
+   */
+  privacy_policy_link_label?: string;
+  /**
+   * Overlay and Privacy Center 'Privacy policy URl'
+   */
+  privacy_policy_url?: string;
+  /**
+   * Overlay 'Privacy preferences link label'
+   */
+  privacy_preferences_link_label?: string;
+  /**
+   * Overlay 'Reject button displayed on the Banner and 'Privacy Preferences' of Privacy Center 'Reject button label'
+   */
   reject_button_label?: string;
-  regions: Array<PrivacyNoticeRegion>;
+  /**
+   * Overlay 'Privacy preferences 'Save' button label
+   */
+  save_button_label?: string;
+  /**
+   * Overlay 'Banner title' or Privacy Center 'title'
+   */
+  title?: string;
+  /**
+   * If None, no edits will be made to regions.  If an empty list, all regions will be removed.
+   */
+  regions?: Array<PrivacyNoticeRegion>;
 };